* superkaleidoscope *
- This is the commit where we break out the kaleidoscope into a separate
  file because we about to do some crazy shizz up in this mofo.

- That is how the commit was granted its name which was fixed when this
  line was added.

- Started sending lists to reduce arguments in the client.
