+++ README.md
@@ -8,7 +8,7 @@ Requires the following:
 
 - python3
 - unicornhat python module (https://github.com/pimoroni/unicorn-hat)
-- ansicolors python module (https://pypi.org/project/ansicolor/)
+- ansicolors python module (https://github.com/numerodix/ansicolor)
 
 ## Documentation
