+++ uhl/lsd.py
@@ -4,7 +4,6 @@
 import time
 import random
 import unicornhat
-from . import lsc
 
 def initunicornhat():
     """ This function initialises the unicornhat hat. """
@@ -37,30 +36,3 @@ def pulse(myred, mygreen, myblue):
     unicornhat.set_all(0, 0, 0)
     unicornhat.show()
     time.sleep(0.5)
-
-# Kscope
-
-def kscope(width, height , xcoord, ycoord, myred, mygreen, myblue, mytime):
-    """ Freakin' Kaleidoscope F.X. """
-    choices = [0, 1]
-    choice = random.choice(choices)
-    redshift, greenshift, blueshift = lsc.getshift()
-    for i in range(mytime):
-        if width == height:
-            for thisrot in range(0,360,90):
-                unicornhat.rotation(thisrot)
-                unicornhat.set_pixel(xcoord, ycoord, myred, mygreen, myblue)
-            unicornhat.show()
-        else:
-            for thisrot in range(0,270,180):
-                unicornhat.rotation(thisrot)
-                unicornhat.set_pixel(xcoord, ycoord, myred, mygreen, myblue)
-            unicornhat.show()
-        if choice == 0:
-            myred, mygreen, myblue = lsc.warpcolour(myred, mygreen, myblue)
-        else:
-            myred, mygreen, myblue = lsc.shiftcolour(myred, mygreen, myblue, redshift, blueshift, greenshift)
-        time.sleep(1)
-        xcoord, ycoord = getcoords(width,height)
-    unicornhat.set_all(0, 0, 0)
-    unicornhat.show()
